+++ lib/src/core/ec_dispose_less_element.dart
@@ -4,18 +4,18 @@ import 'package:flutter/cupertino.dart';
 class ECDisposeElement<T> extends StatelessElement {
   Function()? dispose;
   Function()? afterBulder;
-  Function(Function())? ec_update;
+  Function(Function())? ecUpdate;
 
   ECDisposeElement(StatelessWidget widget,
-      {T Function()? builderVM, this.dispose, this.afterBulder, this.ec_update})
+      {T Function()? builderVM, this.dispose, this.afterBulder, this.ecUpdate})
       : super(widget) {
     if (builderVM != null) {}
     if (afterBulder != null) {
       afterBulder!();
     }
 
-    if (ec_update != null) {
-      ec_update!(markNeedsBuild);
+    if (ecUpdate != null) {
+      ecUpdate!(markNeedsBuild);
     }
   }