+++ lib/src/core/ec_form_config.dart
@@ -1,5 +1,4 @@
-import 'package:flutter/material.dart' show Colors, Color;
-
+import 'package:flutter/material.dart' show Color;
 import 'ec_color.dart';
 
 class ECCormConfig {
@@ -7,8 +6,6 @@ class ECCormConfig {
 
   /// 表单组件与两边space，最终已padding方式呈现
   static double space = 16;
-
- 
 
   static setup({required double horizontalSpace, required Color selectColor}) {
     space = horizontalSpace;